pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"
[tool.setuptools]
package-dir = {"" = "LVLMRecovery/src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["LVLMRecovery/src"]

[project]
name = "LVLMRecovery"
version = "0.1.0"
description = "Autonomous drone landing system based on LVLMs"
readme = "README.md"
license = { text = "MIT License" }
authors = [
  { name="Diego Ortiz"},
  { name="Mohit Agrawal"},
  { name="Yash Malegaonkar"},
  { name="Luis Burbano"},
  { name="Axel Anderson"},
  { name="Gy¨orgy D´an"},
  { name="Henrik Sandberg"},
  { name="Alvaro A. Cardenas"}
]
requires-python = ">=3.10"
dependencies = [
  "cosysairsim ~= 3.3.0",
  "matplotlib ~= 3.10.3",
  "numpy >= 1.26.4",
  "opencv-python >= 4.11",
  "openai >= 1.93.0",
  "open3d ~= 0.19.0",
  "pandas ~= 2.3.0",
  "pillow ~= 11.2.1",
  "pynput >= 1.8",
  "scikit-learn ~= 1.7.0",
  "transformers",
  "torch"
]
